// src/store/cookie-bookmarks/cookie-bookmarks.types.ts
export enum COOKIE_BOOKMARKS_ACTION_TYPES {
   FETCH_BOOKMARKS_START = 'cookie-bookmarks/FETCH_BOOKMARKS_START',
   FETCH_BOOKMARKS_SUCCESS = 'cookie-bookmarks/FETCH_BOOKMARKS_SUCCESS',
   FETCH_BOOKMARKS_FAILED = 'cookie-bookmarks/FETCH_BOOKMARKS_FAILED',
   REFRESH_BOOKMARKS_START = 'cookie-bookmarks/REFRESH_BOOKMARKS_START',
   REFRESH_BOOKMARKS_SUCCESS = 'cookie-bookmarks/REFRESH_BOOKMARKS_SUCCESS',
   REFRESH_BOOKMARKS_FAILED = 'cookie-bookmarks/REFRESH_BOOKMARKS_FAILED',
}

export enum ADD_COOKIE_BOOKMARK_ACTION_TYPES {
   ADD_BOOKMARK_START = 'cookie-bookmarks/ADD_BOOKMARK_START',
   ADD_BOOKMARK_SUCCESS = 'cookie-bookmarks/ADD_BOOKMARK_SUCCESS',
   ADD_BOOKMARK_FAILED = 'cookie-bookmarks/ADD_BOOKMARK_FAILED',
}

export enum REMOVE_COOKIE_BOOKMARKS_ACTION_TYPES {
   REMOVE_BOOKMARK_START = 'cookie-bookmarks/REMOVE_BOOKMARK_START',
   REMOVE_BOOKMARK_SUCCESS = 'cookie-bookmarks/REMOVE_BOOKMARK_SUCCESS',
   REMOVE_BOOKMARK_FAILED = 'cookie-bookmarks/REMOVE_BOOKMARK_FAILED',
}